ah what a piece of-

+++ commands/games.js
@@ -981,7 +981,7 @@ module.exports.peasants.chess = async (message, content, lang, i18n, OpalBot) =>
         if (bot_message.deletable) {
             bot_message.delete();
         }
-        if (message.deletable) {
+        if (message.deletable && typeof message.delete == 'function') {
             setTimeout(() => message.delete(), 500);
         }
     }
